Pass allocatorCapacity as Long, convert to integer in the ldrc layer

+++ extensions-core/multi-stage-query/src/main/java/org/apache/druid/msq/querykit/WindowOperatorQueryFrameProcessor.java
@@ -320,7 +320,7 @@ public class WindowOperatorQueryFrameProcessor implements FrameProcessor<Object>
         OffsetLimit.limit(Integer.MAX_VALUE),
         null,
         null,
-        (int) frameWriterFactory.allocatorCapacity()
+        frameWriterFactory.allocatorCapacity()
     );
     return ldrc;
   }
